Send 24 bytes instead of 4bytes x 6. Added debugging capabilites.
The Arduino now sends one frame of 24 bytes rather than six frames of 4 bytes. Added some safety measures by defining the number of encoders at the top and then using sizeof to find the size of the count variable so that we send the exact right number of bytes every time. Also added debugging capabilities. All you have to do is uncomment debug to use it.